@@ -3,10 +3,11 @@ package co.aisaac.scaper;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Transient;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Transient;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -15,18 +16,18 @@ import java.util.List;
 @Entity(name = "jobs")
 public class Job {
 
-	@Id
-	private Long id;
+    @Id
+    private Long id;
 
-	String title;
-	String url;
-	String company;
-	@Transient
-	List<String> companyStatuses = new ArrayList<>();
-	String subtitle;
-	String description;
-	String status;
+    String title;
+    String url;
+    String company;
+    @Transient
+    List<String> companyStatuses = new ArrayList<>();
+    String subtitle;
+    String description;
+    String status;
 
-	@Column(name = "search_term")
-	String searchTerm;
+    @Column(name = "search_term")
+    String searchTerm;
 }
