@@ -22,152 +22,163 @@ import java.util.regex.Pattern;
 @Controller
 public class WebController {
 
-    private final JobRepo repo;
+  private final JobRepo repo;
 
-    private Filter filter;
+  private Filter filter;
 
-    public WebController(JobRepo repo) {
-        this.repo = repo;
-        this.filter = new Filter();
+
+  public WebController(JobRepo repo) {
+    this.repo = repo;
+    this.filter = new Filter();
+  }
+
+
+  @GetMapping("/")
+  public String all(Model model) {
+    List<Job> jobs = repo.findAll();
+
+    augmentCompanies(jobs);
+
+    jobs = filterJobs(filter, jobs);
+
+    model.addAttribute("jobs", jobs);
+    model.addAttribute("filter", filter);
+    return "jobs";
+  }
+
+
+  @PostMapping("/")
+  public String allFiltered(Filter filter) {
+    this.filter = filter;
+    return "redirect:/";
+  }
+
+
+  @ResponseBody
+  @PostMapping("/status/{id}/{status}")
+  public ResponseEntity<String> updateStatus(@PathVariable String id, @PathVariable String status) {
+    Long idd = Long.parseLong(id);
+    Optional<Job> optJob = repo.findById(idd);
+
+    if (optJob.isEmpty())
+      return ResponseEntity.notFound().build();
+
+    Job job = optJob.get();
+    job.setStatus(status);
+
+    repo.save(job);
+    return ResponseEntity.ok().build();
+  }
+
+
+  private List<Job> filterJobs(Filter filter, List<Job> all) {
+
+    // filter status
+    List<String> statuses = filter.getStatuses();
+    List<Job> filtered = new ArrayList<>(all);
+    if (!statuses.isEmpty()) {
+      filtered = filtered.stream().filter(job -> {
+        for (String status : statuses)
+          if (job.status.trim().equalsIgnoreCase(status.trim()))
+            return true;
+
+        return false;
+      }).toList();
     }
 
-    @GetMapping("/")
-    public String all(Model model) {
-        List<Job> jobs = repo.findAll();
+    // filter company
+    List<String> companies =
+        Arrays.stream(filter.companySearch.split(",")).filter(s -> !s.isBlank()).toList();
+    if (!companies.isEmpty()) {
+      filtered = filtered.stream().filter(job -> {
+        for (String company : companies)
+          if (job.company.trim().toLowerCase(Locale.ROOT)
+              .contains(company.trim().toLowerCase(Locale.ROOT)))
+            return true;
 
-        augmentCompanies(jobs);
-
-        jobs = filterJobs(filter, jobs);
-
-        model.addAttribute("jobs", jobs);
-        model.addAttribute("filter", filter);
-        return "jobs";
+        return false;
+      }).toList();
     }
 
-    @PostMapping("/")
-    public String allFiltered(Filter filter) {
-        this.filter = filter;
-        return "redirect:/";
+
+    // filter search term
+    List<String> searchTerms =
+        Arrays.stream(filter.searchTerms.split(",")).filter(s -> !s.isBlank()).toList();
+    if (!searchTerms.isEmpty()) {
+      filtered = filtered.stream().filter(job -> {
+        var desc = job.description.toLowerCase();
+        for (String term : searchTerms) {
+          if (desc.contains(term.trim().toLowerCase()))
+            return true;
+          if (job.title.toLowerCase(Locale.ROOT).contains(term.trim().toLowerCase()))
+            return true;
+          if (job.subtitle.toLowerCase(Locale.ROOT).contains(term.trim().toLowerCase()))
+            return true;
+        }
+        return false;
+      }).toList();
     }
 
-    @ResponseBody
-    @PostMapping("/status/{id}/{status}")
-    public ResponseEntity<String> updateStatus(@PathVariable String id, @PathVariable String status) {
-        Long idd = Long.parseLong(id);
-        Optional<Job> optJob = repo.findById(idd);
-        if (optJob.isEmpty())
-            return ResponseEntity.notFound().build();
-
-        Job job = optJob.get();
-        job.setStatus(status);
-        repo.save(job);
-        return ResponseEntity.ok().build();
+    // filter title
+    List<String> titleSearchTerms =
+        Arrays.stream(filter.titleSearch.split(",")).filter(s -> !s.isBlank()).toList();
+    if (!titleSearchTerms.isEmpty()) {
+      filtered = filtered.stream().filter(job -> {
+        for (String term : titleSearchTerms) {
+          if (job.title.toLowerCase(Locale.ROOT).contains(term.trim().toLowerCase()))
+            return true;
+        }
+        return false;
+      }).toList();
     }
 
-    private List<Job> filterJobs(Filter filter, List<Job> all) {
 
-        // filter status
-        List<String> statuses = filter.getStatuses();
-        List<Job> filtered = new ArrayList<>(all);
-        if (!statuses.isEmpty()) {
-            filtered = filtered.stream().filter(job -> {
-                for (String status : statuses)
-                    if (job.status.trim().equalsIgnoreCase(status.trim()))
-                        return true;
+    // highlight search term
+    highlight(searchTerms, filtered);
 
-                return false;
-            }).toList();
-        }
+    // sort
+    filtered =
+        filtered.stream().sorted((o1, o2) -> o1.company.compareToIgnoreCase(o2.company)).toList();
 
-        // filter company
-        List<String> companies = Arrays.stream(filter.companySearch.split(",")).filter(s -> !s.isBlank()).toList();
-        if (!companies.isEmpty()) {
-            filtered = filtered.stream().filter(job -> {
-                for (String company : companies)
-                    if (job.company.trim().toLowerCase(Locale.ROOT).contains(company.trim().toLowerCase(Locale.ROOT)))
-                        return true;
-
-                return false;
-            }).toList();
-        }
+    return filtered;
+  }
 
 
-        // filter search term
-        List<String> searchTerms = Arrays.stream(filter.searchTerms.split(",")).filter(s -> !s.isBlank()).toList();
-        if (!searchTerms.isEmpty()) {
-            filtered = filtered.stream().filter(job -> {
-                var desc = job.description.toLowerCase();
-                for (String term : searchTerms) {
-                    if (desc.contains(term.trim().toLowerCase()))
-                        return true;
-                    if (job.title.toLowerCase(Locale.ROOT).contains(term.trim().toLowerCase()))
-                        return true;
-                    if (job.subtitle.toLowerCase(Locale.ROOT).contains(term.trim().toLowerCase()))
-                        return true;
-                }
-                return false;
-            }).toList();
-        }
+  private void augmentCompanies(List<Job> jobs) {
+    // add statuses to each company
 
-        // filter title
-        List<String> titleSearchTerms = Arrays.stream(filter.titleSearch.split(",")).filter(s -> !s.isBlank()).toList();
-        if (!titleSearchTerms.isEmpty()) {
-            filtered = filtered.stream().filter(job -> {
-                for (String term : titleSearchTerms) {
-                    if (job.title.toLowerCase(Locale.ROOT).contains(term.trim().toLowerCase()))
-                        return true;
-                }
-                return false;
-            }).toList();
-        }
+    Map<String, Set<String>> comps = new HashMap<>();
 
-
-        // highlight search term
-        highlight(searchTerms, filtered);
-
-        // sort
-        filtered = filtered.stream()
-                .sorted((o1, o2) -> o1.company.compareToIgnoreCase(o2.company))
-                .toList();
-
-        return filtered;
+    for (Job job : jobs) {
+      var comp = job.company;
+      if (comps.containsKey(comp)) {
+        comps.get(comp).add(job.status);
+      } else {
+        comps.put(comp, new HashSet<>());
+        comps.get(comp).add(job.status);
+      }
     }
 
-    private void augmentCompanies(List<Job> jobs) {
-        // add statuses to each company
-
-        Map<String, Set<String>> comps = new HashMap<>();
-
-        for (Job job : jobs) {
-            var comp = job.company;
-            var status = job.status;
-            if (comps.containsKey(comp)) {
-                comps.get(comp).add(status);
-            } else {
-                comps.put(comp, new HashSet<>());
-                comps.get(comp).add(status);
-            }
-        }
-
-        for (Job job : jobs) {
-            var statuses = comps.get(job.company);
-            job.companyStatuses = statuses.stream().toList();
-        }
+    for (Job job : jobs) {
+      var statuses = comps.get(job.company);
+      job.companyStatuses = statuses.stream().toList();
     }
+  }
 
-    private void highlight(List<String> searchTerms, List<Job> filtered) {
-        if (searchTerms.isEmpty())
-            return;
-        if (filtered.isEmpty())
-            return;
 
-        for (String searchTerm : searchTerms) {
-            String pattern = "(?i)" + Pattern.quote(searchTerm.trim());
-            String st = "<span class=\"highlight\">" + searchTerm + "</span>";
-            // for each search term
-            for (Job job : filtered) {
-                job.description = job.description.replaceAll(pattern, st);
-            }
-        }
+  private void highlight(List<String> searchTerms, List<Job> filtered) {
+    if (searchTerms.isEmpty())
+      return;
+    if (filtered.isEmpty())
+      return;
+
+    for (String searchTerm : searchTerms) {
+      String pattern = "(?i)" + Pattern.quote(searchTerm.trim());
+      String st = "<span class=\"highlight\">" + searchTerm + "</span>";
+      // for each search term
+      for (Job job : filtered) {
+        job.description = job.description.replaceAll(pattern, st);
+      }
     }
+  }
 }
